Add support for kubernetes jobs
Added a new issue type to the adapter descriptor to support Jobs. A job is created and then a message is sent to Kubernetes to start the job, and the rest of the job is written back to the requesting endpoint. An example job is included (test.json), and a Docker file that will be executed as the job.